+++ actions.js
@@ -352,7 +352,7 @@ export const actions = {
                     global.evolution['final'] = 100;
                     global.evolution['sentience'] = { count: 0 };
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         if (global.evolution['chitin']){
                             global.evolution['sporgar'] = { count: 0 };
                             global.evolution['shroomi'] = { count: 0 };
@@ -398,7 +398,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['mantis'] = { count: 0 };
                         global.evolution['scorpid'] = { count: 0 };
                         global.evolution['antid'] = { count: 0 };
@@ -478,7 +478,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['human'] = { count: 0 };
                         global.evolution['orc'] = { count: 0 };
                         global.evolution['elven'] = { count: 0 };
@@ -520,7 +520,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['troll'] = { count: 0 };
                         global.evolution['orge'] = { count: 0 };
                         global.evolution['cyclops'] = { count: 0 };
@@ -562,7 +562,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['kobold'] = { count: 0 };
                         global.evolution['goblin'] = { count: 0 };
                         global.evolution['gnome'] = { count: 0 };
@@ -604,7 +604,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['cath'] = { count: 0 };
                         global.evolution['wolven'] = { count: 0 };
                         global.evolution['centaur'] = { count: 0 };
@@ -644,7 +644,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['balorg'] = { count: 0 };
                         global.evolution['imp'] = { count: 0 };
                         addAction('evolution','balorg');
@@ -680,7 +680,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['sharkin'] = { count: 0 };
                         global.evolution['octigoran'] = { count: 0 };
                         addAction('evolution','sharkin');
@@ -742,7 +742,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['arraak'] = { count: 0 };
                         global.evolution['pterodacti'] = { count: 0 };
                         global.evolution['dracnid'] = { count: 0 };
@@ -776,7 +776,7 @@ export const actions = {
                     global.evolution['sentience'] = { count: 0 };
                     global.evolution['final'] = 100;
                     addAction('evolution','sentience');
-                    if (global.race.seeded){
+                    if (global.race.seeded || global.stats.achieve['creator']){
                         global.evolution['tortoisan'] = { count: 0 };
                         global.evolution['gecko'] = { count: 0 };
                         global.evolution['slitheryn'] = { count: 0 };
@@ -6913,7 +6913,6 @@ export const actions = {
             title: loc('tech_fanaticism'),
             desc: loc('tech_fanaticism'),
             reqs: { theology: 2 },
-            not_trait: ['herbivore'],
             grant: ['theology',3],
             cost: {
                 Knowledge(){ return 2500; }
@@ -7884,7 +7883,7 @@ export const actions = {
                     k_inc *= 1.015;
                 }
                 plasmid = challenge_multiplier(plasmid);
-                let phage = Math.floor(Math.log2(plasmid) * Math.E);
+                let phage = challenge_multiplier(Math.floor(Math.log2(plasmid) * Math.E));
                 return `<div>${loc('star_dock_genesis_effect1')}</div><div class="has-text-special">${loc('star_dock_genesis_effect2',[plasmid])}</div><div class="has-text-special">${loc('star_dock_genesis_effect3',[phage])}</div>`;
             },
             action(){
@@ -9235,31 +9234,37 @@ function fanaticism(god){
             fanaticTrait('brute');
             break;
         case 'cath':
-            fanaticTrait('carnivore');
-            if (global.tech['farm'] >= 1){
-                global.tech['hunting'] = 2;
+            if (global.race['herbivore']){
+                randomMinorTrait();
+                arpa('Genetics');
             }
-            else if (global.tech['agriculture'] >= 3){
-                global.tech['hunting'] = 1;
-            }
-            if (global.city['farm']){
-                global.city['lodge'] = { count: global.city.farm.count };
-                delete global.city['farm'];
-            }
-            if (global.city['silo']){
-                global.city['smokehouse'] = { count: global.city.silo.count };
-                delete global.city['silo'];
-            }
-            if (global.city['mill']){
-                delete global.city['mill'];
-            }
-            delete global.tech['agriculture'];
-            delete global.tech['farm'];
-            global.civic.farmer.workers = 0;
-            global.civic.farmer.max = 0;
-            global.civic.farmer.display = false;
-            if (global.race.species === 'entish'){
-                unlockAchieve(`madagascar_tree`);
+            else {
+                fanaticTrait('carnivore');
+                if (global.tech['farm'] >= 1){
+                    global.tech['hunting'] = 2;
+                }
+                else if (global.tech['agriculture'] >= 3){
+                    global.tech['hunting'] = 1;
+                }
+                if (global.city['farm']){
+                    global.city['lodge'] = { count: global.city.farm.count };
+                    delete global.city['farm'];
+                }
+                if (global.city['silo']){
+                    global.city['smokehouse'] = { count: global.city.silo.count };
+                    delete global.city['silo'];
+                }
+                if (global.city['mill']){
+                    delete global.city['mill'];
+                }
+                delete global.tech['agriculture'];
+                delete global.tech['farm'];
+                global.civic.farmer.workers = 0;
+                global.civic.farmer.max = 0;
+                global.civic.farmer.display = false;
+                if (global.race.species === 'entish'){
+                    unlockAchieve(`madagascar_tree`);
+                }
             }
             break;
         case 'wolven':
@@ -9368,7 +9373,7 @@ function fanaticism(god){
 function fanaticTrait(trait){
     if (global.race[trait]){
         randomMinorTrait();
-        arpa('Genetics')
+        arpa('Genetics');
     }
     else {
         global.race[trait] = 1;
@@ -9400,7 +9405,7 @@ function bioseed(){
     }
     new_plasmid = challenge_multiplier(new_plasmid);
     plasmid += new_plasmid;
-    let new_phage = Math.floor(Math.log2(new_plasmid) * Math.E);
+    let new_phage = challenge_multiplier(Math.floor(Math.log2(new_plasmid) * Math.E));
     phage += new_phage;
     global.stats.reset++;
     global.stats.tdays += global.stats.days;
@@ -9416,6 +9421,10 @@ function bioseed(){
     unlockAchieve(`seeder`);
     let new_biome = unlockAchieve(`biome_${biome}`);
     let new_genus = unlockAchieve(`genus_${genus}`);
+    let probes = global.starDock.probes.count + 1;
+    if (global.stats.achieve['explorer']){
+        probes += global.stats.achieve['explorer'];
+    }
     global['race'] = { 
         species : 'protoplasm', 
         gods: god,
@@ -9423,7 +9432,7 @@ function bioseed(){
         Plasmid: { count: plasmid },
         Phage: { count: phage },
         seeded: true,
-        probes: global.starDock.probes.count + 1,
+        probes: probes,
         seed: Math.floor(Math.seededRandom(10000)),
     };
     global.city = {